Corrección de errores básicos

+++ Amazonia/Model/DAOs/ProductDao/IProductDao.cs
@@ -7,7 +7,7 @@ namespace Es.Udc.DotNet.Amazonia.Model.DAOs.ProductDao
 {
     public interface IProductDao : IGenericDao<Product, Int64>
     {
-        List<ProductDTO> FindByKeyWordAndCategory(string keyWord, int categoryId);
+        List<ProductDTO> FindByKeyWordAndCategory(string keyWord, long categoryId);
 
         List<ProductDTO> FindByKeyWord(string keyWord);
     }
